DrinkList refactor and drink name and description editable in UI, but not saving to db yet
