+++ src/EditBox.js
@@ -3,14 +3,13 @@ import React from "react";
 export default function EditBox({
   edit,
   handleEdit,
-  id,
-  type
+  id
 }){
     return (    
       <input
           id={id}
           size={edit.length}
-          type={type}
+          type='text'
           value={edit}
           onChange={handleEdit}
       />
